requirements.txt
django
psycopg2
django-crispy-forms
pillow
reportlab
djangorestframework